changed the port number again and in product.controller.js file added a condition for delete function for user if add wrong id to delete a product

+++ product-details-project/backend/controllers/product.controller.js
@@ -70,11 +70,15 @@ export const deleteProduct = async (req, res) => {
     const { id } = req.params;
     console.log("id:", id);
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id!" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted!" });
     } catch (error) {
         console.log("error in deleting product:", error.message);
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Internal Server Error" });
     }
 };
